q-and-a-orchestra-agent/requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# Async
asyncio==3.4.3
aiohttp==3.9.0

# Database
psycopg[binary]==3.1.12  # PostgreSQL
SQLAlchemy==2.0.23
alembic==1.13.0

# Time-series DB
timescaledb==5.0.0

# Vector DB
weaviate-client==4.1.1  # or pinecone-client==3.0.0

# LLM/AI
openai==1.3.0
anthropic==0.12.0
ollama==0.1.0

# Embeddings (lightweight)
sentence-transformers==2.2.2

# NLP/Validation
transformers==4.36.0
torch==2.1.0

# Metrics/Observability
prometheus-client==0.19.0
python-json-logger==2.0.7
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0

# Caching
redis==5.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0

# Code quality
black==23.12.0
pylint==3.0.0
mypy==1.7.0

# DevOps
docker==7.0.0
kubernetes==28.0.0

# Utils
pyyaml==6.0.1
requests==2.31.0
httpx==0.25.0
uuid==1.30
tiktoken==0.5.2
